requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2
langsmith>=0.0.87
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
pytest>=8.0.0 